[TFP-535] Move several functions into different files
