+++ cli/app/access_nix.go
@@ -18,11 +18,8 @@ package app
 
 import "golang.org/x/sys/unix"
 
-// Read/Write/Execute Constants
 const (
-	ACCESS_R_OK = unix.R_OK
 	ACCESS_W_OK = unix.W_OK
-	ACCESS_X_OK = unix.X_OK
 )
 
 func checkAccess(path string, perm uint32) error {
