+++ pkg/tools/util.go
@@ -16,18 +16,14 @@ package tools
 
 import (
 	"fmt"
-	"github.com/akamai/cli/pkg/version"
-	"io"
-	"os"
-	"os/exec"
-	"path/filepath"
-	"strings"
-	"syscall"
-
 	akamai "github.com/akamai/cli-common-golang"
+	"github.com/akamai/cli/pkg/version"
 	"github.com/fatih/color"
 	"github.com/mitchellh/go-homedir"
 	"github.com/urfave/cli"
+	"os"
+	"path/filepath"
+	"strings"
 )
 
 func Self() string {
@@ -57,25 +53,6 @@ func GetAkamaiCliSrcPath() (string, error) {
 	cliHome, _ := GetAkamaiCliPath()
 
 	return filepath.Join(cliHome, "src"), nil
-}
-
-func PassthruCommand(executable []string) error {
-	subCmd := exec.Command(executable[0], executable[1:]...)
-	subCmd.Stdin = os.Stdin
-	subCmd.Stderr = os.Stderr
-	subCmd.Stdout = os.Stdout
-	err := subCmd.Run()
-
-	exitCode := 1
-	if exitError, ok := err.(*exec.ExitError); ok {
-		if waitStatus, ok := exitError.Sys().(syscall.WaitStatus); ok {
-			exitCode = waitStatus.ExitStatus()
-		}
-	}
-	if err != nil {
-		return cli.NewExitError("", exitCode)
-	}
-	return nil
 }
 
 func Githubize(repo string) string {
@@ -109,41 +86,4 @@ func ShowBanner() {
 	fmt.Fprintf(akamai.App.ErrWriter, fg.Sprintf(ws+title+ws+"\n"))
 	fmt.Fprintf(akamai.App.ErrWriter, bg.Sprintf(strings.Repeat(" ", 60)+"\n"))
 	fmt.Fprintln(akamai.App.ErrWriter)
-}
-
-// We must copy+unlink the file because moving files is broken across filesystems
-func MoveFile(src string, dst string) error {
-	sourceFileStat, err := os.Stat(src)
-	if err != nil {
-		return err
-	}
-
-	if !sourceFileStat.Mode().IsRegular() {
-		return fmt.Errorf("%s is not a regular file", src)
-	}
-
-	source, err := os.Open(src)
-	if err != nil {
-		return err
-	}
-	defer source.Close()
-
-	destination, err := os.Create(dst)
-	if err != nil {
-		return err
-	}
-	defer destination.Close()
-	_, err = io.Copy(destination, source)
-
-	if err != nil {
-		return err
-	}
-
-	err = os.Chmod(dst, 0755)
-	if err != nil {
-		return err
-	}
-
-	err = os.Remove(src)
-	return err
 }
